tests/wpt/tests/html/syntax/charset/meta-commented-out.html

<!DOCTYPE html>
<html>
  <head>
      <!--<meta charset="ISO-8859-8">-->
      <script src="/resources/testharness.js"></script>
      <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      "use strict";
      test(() => {
        assert_not_equals(document.characterSet, "ISO-8859-8");
      }, "document.characterSet must not be determined by commented meta charset");
    </script>
  </body>
</html>
